Create separate adapters for OSF models
- Instead of making the application adapter host
  the collections API, make it the host for just the
  adapters for collection models (collection, group and
  item).
- Add node, preprint and registration adapters with
  staging-api.osf.io as the host, that sends the bearer
  token with each request. Note that ajax withCredentials
  must be false!
