+++ client/app/adapters/collection.js
@@ -1,8 +1,18 @@
-import Application from './application';
+import Ember from 'ember';
+import DS from 'ember-data';
 
-export default Application.extend({
+
+export default DS.JSONAPIAdapter.extend({
+    session: Ember.inject.service(),
+    host: 'http://localhost:8000',
+    namespace: 'api',
+    ajax(url, method, hash) {
+        hash = hash || {};
+        hash.crossOrigin = true;
+        hash.xhrFields = { withCredentials: true };
+        return this._super(url, method, hash);
+    },
     buildURL() {
-        // Embed linked_nodes
         var base = this._super(...arguments);
         return `${base}/`;
     }
