+++ client/app/adapters/registration.js
@@ -1,0 +1,4 @@
+import NodeAdapter from './node';
+
+export default NodeAdapter.extend({
+});
